Make forwardToKubeDNS work in the NodeLocal DNSCache template
This fixes the rendering of the Corefile of the NodeLocal DNSCache
template when setting forwardToKubeDNS to false (or not setting it).

Previously, due to not dereferencing the bool pointer, the Corefile was
always rendered with the true clause, due to checking the address
instead of the actual value of the variable.

With this fix, the templating mechanism will actually respect the value
of forwardToKubeDNS and render it accordingly.