Add nodeid to older cinder-csi-plugin deployments
This will add the `--nodeid` flag to older cinder-csi-plugin
deployments, because the plugin requires it up to version 1.21.

The approach naively assumes that a user has not overridden the
cinder-csi-plugin image and thus uses the Kubernetes version to
determine if the nodeid flag needs to be set.

It is recommended anyway to keep the minor versions of the
cinder-csi-plugin and Kubernetes the same.