WIP OpenStack: Use task engine to retry failed servers
This will make use of the kOps taks engine to retry failed servers.

The former approach <PR link> had the side effect of not making kOps
fail when the last retry failed: When the instance task got retried by
the task engine it successfully reconciled the port/floating ip, as
there was now an existing/actual - although erroneous - server present.